rules: add extra_check! macro for conditionally compiling rule checks
Add extra_check! macro, and the extra-rule-checks feature, allowing
rules to specify potentially expensive extra checks to be optionally
ran.

In the future, we will assume that the highest priority applicable rule
will be applied, regardless of where in the tree it applies. However, we
currently only apply the most applicable rule for our current
expression, ignoring any higher priority rules that could apply to child
expressions.

For example, consider partial evaluation. In general, we want to ensure
that the entire tree is in a simplest form before proceeding.

```
R1 (priority 10): collapse constants in addition
R2 (priority 5): f(M) ~> g(M) /\ h(M) /\ i(M) /\ ...
  where M is a sum
```

For some expression `f(1+b+2)` we would want to apply R1 to the
arguments before doing R2; however, the current rewriter would apply R2
first, as this is the largest expression.

To fix this in the current rewriter, R2 would have to say that it is
only applicable when R1 isn't. This is an expensive check, and would be
unnecessary with the new rewriter; however, it is necessary for now.

extra_check! delineates these checks and will allows us to trivially
compile them out once we have the new rewriter functionality.

These checks may be helpful to retain to identify incorrect rule
priorities or spot bugs in the rewriting engine; hence I suggest keeping
them accessible under a compilation flag in the future.

This work will help my upcoming work on flattening rules.